Created a folder 'section' for Section-based components.
Originally, I removed it because the section and components were similar by function back then (before I created a this new repo). Now however, sections, for me at least, are components that has the same information throughout the website. The PreviewSection is one of them.

PreviewSection is a component that hold these PagePreviews which will be used throughout the website.

Planning to rename the SpotlightSection for a better fitting name,