Change TextSection to TextContainer
To differentiate between TextSection and PagePreviewSection

Sections (in my case) are sections of the page that will be reused but cannot be changed directly.
Containers are components that can be reused and be added with children or additional properties.

This is my structure and is probably not the usual convention.