Implemented the ContactForm
It finally works. I hope this time? It'll clear the form when submitted successfully, and will tell the user of whatever that has happened.

Additionally, I made it so that the button, textarea, and input styling made in ContactForm to be in the index.css to be used globally. That might change, but for now it works. I'll think about it further as time goes on.